+++ NEWS.md
@@ -59,7 +59,7 @@ New library functions
 
 * `hardlink(src, dst)` can be used to create hard links. ([#41639])
 * `diskstat(path=pwd())` can be used to return statistics about the disk. ([#42248])
-* `Iterators.unfoldr(f, init)` can be used to quickly write iterators or as a kind of lazy `while` loop ([#43203])
+* `Iterators.unfold(f, init)` can be used to quickly write iterators or as a kind of lazy `while` loop ([#43203])
 
 New library features
 --------------------